requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
contextlib2==0.6.0.post1
cryptography==3.1.1
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.12.1
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.22.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
multidict==4.7.6
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
rsa==4.6
schema==0.7.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
yarl==1.6.0
